+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.0] - 2021-09-16
+
+### Changed
+
+- Compatible with chaos mesh 2.x. **(After 2.0.x, will start with 2.1.x)**
+- Bump the minimal grafana version to 7.0.0
+- Bump grafana/toolkit to 8.x
+
 ## [0.2.2] - 2021-04-15
 
 ### Changed